preparing experiments

+++ measurements/sessionScripts/terminateSession.sh
@@ -4,5 +4,9 @@
 
 SESSION_NAME=$1
 
-tmux kill-session -t "$SESSION_NAME" 2>/dev/null
+./isRunningSession.sh "$SESSION_NAME"
+if [[ $? -eq 0 ]]; then
+  tmux kill-session -t "$SESSION_NAME" 2>/dev/null
+fi
+
 exit 0
